refactor(import info): split videos & images
